* fix not display 2nd peer

* error login, register

* merge master to dev (#17)

* quick fix

* change peer host

* change peer host

* update peer

* update peer

* update peer

* st

* st

* st

* new domain

* new domain

* new domain

* new domain

* new domain

* new domain

* new apps

* check code

+++ client/src/api/communityService.js
@@ -30,16 +30,11 @@ export const communityUserAdd = async (communityId, user) => {
 };
 
 export const removeUserFromCommunity = async (communityId, userId) => {
-  const res = await axiosClient.delete(
-    `/community/${communityId}/user/${userId}`
-  );
-
+  const res = await axiosClient.delete(`/community/${communityId}/user/${userId}`);
   return res.data;
 };
 
 export const transferCommunityAdmin = async (communityId, userId) => {
-  const res = await axiosClient.put(`/community/${communityId}/user`, {
-    userId,
-  });
+  const res = await axiosClient.put(`/community/${communityId}/user`, {userId});
   return res.data;
 };
